requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.in
#
altgraph==0.17            # via pyinstaller
appdirs==1.4.3
astroid==2.3.3            # via pylint
attrs==19.3.0             # via black
black==19.10b0
bleach==3.1.4             # via readme-renderer
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0              # via cryptography
chardet==3.0.4
click-repl==0.1.6
click==7.1.1
cryptography==2.8         # via secretstorage
docutils==0.16            # via readme-renderer
google-auth==1.14.0
idna==2.9
importlib-metadata==1.6.0  # via keyring, twine
isort==4.3.21             # via pylint
jeepney==0.4.3            # via keyring, secretstorage
jinja2==2.11.1
keyring==21.2.0           # via twine
kubernetes==11.0.0
lazy-object-proxy==1.4.3  # via astroid
markupsafe==1.1.1
mccabe==0.6.1             # via pylint
oauthlib==3.1.0
pathspec==0.7.0           # via black
pip-tools==4.5.1
pkginfo==1.5.0.1          # via twine
prompt-toolkit==3.0.5
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20           # via cffi
pygments==2.6.1           # via readme-renderer
pyinstaller==3.6
pylint==2.4.4
python-dateutil==2.8.1
pyyaml==5.3.1
readme-renderer==25.0     # via twine
regex==2020.2.20          # via black
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1  # via twine
requests==2.23.0
rsa==4.0
secretstorage==3.1.2      # via keyring
six==1.14.0
toml==0.10.0              # via black
tqdm==4.44.1              # via twine
twine==3.1.1
typed-ast==1.4.1          # via astroid, black
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1       # via bleach
websocket-client==0.57.0
wrapt==1.11.2             # via astroid
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
